:sparkles: Add support for long filenames.
Long filenames are capped at ~39 characters (including period) to reduce the number of directory entries required for LFNs. With a limit of 39 characters the maximum number of directory entries required for an LFN entry is four (three for 13 character parts of the name and one for the 8.3 entry with other attributes).